app/requirements.txt
absl-py==0.1.11
aniso8601==3.0.0
astor==0.6.2
backports.weakref==1.0.post1
bleach==1.5.0
click==6.7
enum34==1.1.6
Flask==0.12.2
Flask-RESTful==0.3.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Markdown==2.6.11
MarkupSafe==1.0
marshmallow==2.15.0
mock==2.0.0
numpy==1.14.2
pbr==3.1.1
Pillow==5.0.0
protobuf==3.5.2
python-dotenv==0.8.2
pytz==2018.3
six==1.11.0
tensorboard==1.6.0
tensorflow==1.5.0
tensorflow-serving-api==1.5.0
termcolor==1.1.0
uWSGI==2.0.17
Werkzeug==0.14.1
